backend/requirements.txt
# Core Framework
fastapi==0.121.1
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.44
alembic==1.17.1
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Redis
redis==5.2.1

# Validation & Serialization
pydantic==2.10.4
pydantic-settings==2.7.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# AI Integration (Google Gemini - FREE)
google-genai

# Audio Processing
edge-tts>=6.1.12
pydub==0.25.1

# Logging
structlog==24.4.0

# Development Tools
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
black==24.10.0
ruff==0.8.4
mypy==1.13.0
